requirements_test.txt
-r requirements.txt
pytest
pytest-cov
pytest-django
pytest-pythonpath
